add pass base ut.
